requirements.txt
Flask==2.2.3
flask-cors==3.0.10
python-dotenv
gunicorn